move check_if_semi_supervised
